+++ MatrixSDK/MXError.m
@@ -51,6 +51,7 @@ NSString *const kMXErrCodeStringPasswordNoSymbol        = @"M_PASSWORD_NO_SYMBOL
 NSString *const kMXErrCodeStringPasswordInDictionary    = @"M_PASSWORD_IN_DICTIONARY";
 NSString *const kMXErrCodeStringWeakPassword            = @"M_WEAK_PASSWORD";
 NSString *const kMXErrCodeStringTermsNotSigned          = @"M_TERMS_NOT_SIGNED";
+NSString *const kMXErrCodeStringInvalidPepper           = @"M_INVALID_PEPPER";
 
 NSString *const kMXErrorStringInvalidToken      = @"Invalid token";