+++ MatrixSDK/Utils/MXHTTPClient.m
@@ -78,10 +78,30 @@ NSString* const kMXHTTPClientMatrixErrorNotificationErrorKey = @"kMXHTTPClientMa
      */
     BOOL invalidatedSession;
 }
+
+/**
+ The access token used for authenticated requests.
+ */
+@property (nonatomic, strong) NSString *accessToken;
+
 @end
 
 @implementation MXHTTPClient
 
+#pragma mark - Properties override
+
+// TODO: Set Authorization field only for authenticated requests
+- (void)setAccessToken:(NSString *)accessToken
+{
+    _accessToken = accessToken;
+    
+    [self updateAuthorizationBearHTTPHeaderFieldWithAccessToken:accessToken];
+}
+
+- (NSURL *)baseURL
+{
+    return httpManager.baseURL;
+}
 
 #pragma mark - Public methods
 -(id)initWithBaseURL:(NSString *)baseURL andOnUnrecognizedCertificateBlock:(MXHTTPClientOnUnrecognizedCertificate)onUnrecognizedCertBlock
@@ -112,9 +132,10 @@ NSString* const kMXHTTPClientMatrixErrorNotificationErrorKey = @"kMXHTTPClientMa
         // No need for caching. The sdk caches the data it needs
         [httpManager.requestSerializer setCachePolicy:NSURLRequestReloadIgnoringLocalCacheData];
 
-        // No need for caching. The sdk caches the data it needs
+        // Set authorization HTTP header if access token is present
         if (accessToken)
         {
+            _accessToken = accessToken;
             [httpManager.requestSerializer setValue:[NSString stringWithFormat:@"Bearer %@", accessToken] forHTTPHeaderField:@"Authorization"];
         }
 
@@ -174,6 +195,95 @@ NSString* const kMXHTTPClientMatrixErrorNotificationErrorKey = @"kMXHTTPClientMa
 
     [self tryRequest:mxHTTPOperation method:httpMethod path:path parameters:parameters data:data headers:headers timeout:timeoutInSeconds uploadProgress:uploadProgress success:success failure:failure];
 
+    return mxHTTPOperation;
+}
+
+- (MXHTTPOperation*)requestWithMethod:(NSString *)httpMethod
+                                 path:(NSString *)path
+                           parameters:(NSDictionary*)parameters
+                   needsAuthorization:(BOOL)needsAuthorization
+                              success:(void (^)(NSDictionary *JSONResponse))success
+                              failure:(void (^)(NSError *error))failure
+{
+    return [self requestWithMethod:httpMethod path:path parameters:parameters needsAuthorization:needsAuthorization timeout:-1 success:success failure:failure];
+}
+
+- (MXHTTPOperation*)requestWithMethod:(NSString *)httpMethod
+                                 path:(NSString *)path
+                           parameters:(NSDictionary*)parameters
+                   needsAuthorization:(BOOL)needsAuthorization
+                              timeout:(NSTimeInterval)timeoutInSeconds
+                              success:(void (^)(NSDictionary *JSONResponse))success
+                              failure:(void (^)(NSError *error))failure
+{
+    return [self requestWithMethod:httpMethod path:path parameters:parameters needsAuthorization:needsAuthorization data:nil headers:nil timeout:timeoutInSeconds uploadProgress:nil success:success failure:failure];
+}
+
+
+- (MXHTTPOperation*)requestWithMethod:(NSString *)httpMethod
+                                 path:(NSString *)path
+                           parameters:(NSDictionary*)parameters
+                   needsAuthorization:(BOOL)needsAuthorization
+                                 data:(NSData *)data
+                              headers:(NSDictionary*)headers
+                              timeout:(NSTimeInterval)timeoutInSeconds
+                       uploadProgress:(void (^)(NSProgress *uploadProgress))uploadProgress
+                              success:(void (^)(NSDictionary *JSONResponse))success
+                              failure:(void (^)(NSError *error))failure
+{
+    MXHTTPOperation *mxHTTPOperation = [[MXHTTPOperation alloc] init];
+    
+    [self tryRequest:mxHTTPOperation
+              method:httpMethod
+                path:path
+          parameters:parameters
+                data:data
+            headers:headers
+             timeout:timeoutInSeconds
+      uploadProgress:uploadProgress
+             success:success
+             failure:^(NSError *error) {
+                 
+                 if (needsAuthorization
+                     && error
+                     && self.shouldRenewTokenHandler(error)
+                     && self.renewTokenHandler)
+                 {
+                     dispatch_async(dispatch_get_main_queue(), ^{
+                         
+                         // Remove current access token
+                         self.accessToken = nil;
+                         
+                         mxHTTPOperation.operation = nil;
+                         
+                         typeof(self) __weak weakSelf = self;
+                         
+                         self.renewTokenHandler(^(NSString *accessToken) {
+                             
+                             typeof(self) strongSelf = weakSelf;
+                             
+                             if (strongSelf)
+                             {
+                                 strongSelf.accessToken = accessToken;
+                                 
+                                 [strongSelf tryRequest:mxHTTPOperation
+                                                 method:httpMethod
+                                                   path:path
+                                             parameters:parameters
+                                                   data:data
+                                                headers:headers
+                                                timeout:timeoutInSeconds
+                                         uploadProgress:uploadProgress
+                                                success:success
+                                                failure:failure];
+                             }
+                         }, ^(NSError *error) {
+                             failure(error);
+                         });
+                     });
+                 }
+             }];
+    
     return mxHTTPOperation;
 }
 
@@ -523,6 +633,9 @@ NSString* const kMXHTTPClientMatrixErrorNotificationErrorKey = @"kMXHTTPClientMa
     {
         httpManager.requestSerializer = [AFHTTPRequestSerializer serializer];
     }
+    
+    // Refresh authorization HTTP header field
+    [self updateAuthorizationBearHTTPHeaderFieldWithAccessToken:self.accessToken];
 }
 
 
@@ -790,6 +903,18 @@ NSString* const kMXHTTPClientMatrixErrorNotificationErrorKey = @"kMXHTTPClientMa
     return [[MXError alloc] initWithErrorCode:json[kMXErrorCodeKey]
                                         error:json[kMXErrorMessageKey]
                                      userInfo:mxErrorUserInfo];
+}
+
+- (void)updateAuthorizationBearHTTPHeaderFieldWithAccessToken:(NSString *)accessToken
+{
+    if (accessToken)
+    {
+        [httpManager.requestSerializer setValue:[NSString stringWithFormat:@"Bearer %@", accessToken] forHTTPHeaderField:@"Authorization"];
+    }
+    else
+    {
+        [httpManager.requestSerializer setValue:nil forHTTPHeaderField:@"Authorization"];
+    }
 }
 
 + (void)logRequestFailure:(MXHTTPOperation*)mxHTTPOperation