+++ CHANGES.rst
@@ -3,11 +3,17 @@ Changes in Matrix iOS SDK in 0.13.2 (2019-08-)
 
 Improvements:
  * MXServiceTerms: A class to support MSC2140 (Terms of Service API) (vector-im/riot-ios#2600).
- * MXRestClient: Remove Identity Server URL fallback to homeserver one's when there is no Identity Server configured.
+ * MXRestClient: Remove identity server URL fallback to homeserver one's when there is no identity server configured.
  * MXHTTPClient: Improve M_LIMIT_EXCEEDED error handling: Do not wait to try again if the mentioned delay is too long.
  * MXEventTimeline: The roomEventFilter property is now writable (vector-im/riot-ios#2615).
  * VoIP: Make call start if there is no STUN server.
  * MXMatrixVersions: Add doesServerRequireIdentityServerParam and doesServerAcceptIdentityAccessToken properties.
+ * Create MXIdentityServerRestClient and MXIdentityService to manage identity server requests (vector-im/riot-ios#2647).
+ * MXHTTPClient: Add access token renewal plus request retry mechanism.
+ * MXIdentityService: Support identity server v2 API. Handle identity server v2 API authentification and use the hashed v2 lookup API for 3PIDs (vector-im/riot-ios#2603 and /vector-im/riot-ios#2652).
+
+API break:
+ * MXRestClient: Remove identity server requests. Now MXIdentityService is used to perform identity server requests.
 
 Changes in Matrix iOS SDK in 0.13.1 (2019-08-08)
 ===============================================